requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2 ; python_version >= '3.7'
asttokens==2.4.1
beautifulsoup4==4.12.2 ; python_full_version >= '3.6.0'
bleach==6.1.0 ; python_version >= '3.8'
bs4==0.0.1
certifi==2023.11.17 ; python_version >= '3.6'
charset-normalizer==3.3.2 ; python_full_version >= '3.7.0'
colorama==0.4.6 ; sys_platform == 'win32'
crispy-bootstrap5==2023.10
decorator==5.1.1 ; python_version >= '3.5'
django==4.2.7
django-bootstrap5==23.3
django-browser-reload==1.12.1
django-crispy-forms==2.1
django-floppyforms==1.9.0
django-mdeditor==0.1.20
django-model-utils==4.3.1 ; python_version >= '3.7'
django-notifications-hq==1.8.3
django-tailwind[reload]==3.6.0
django-widget-tweaks==1.5.0
executing==2.0.1 ; python_version >= '3.5'
fuzzychinese==0.1.5
idna==3.6 ; python_version >= '3.5'
importlib-metadata==4.8.3 ; python_version >= '3.6'
ipython==8.18.1 ; python_version >= '3.9'
jedi==0.19.1 ; python_version >= '3.6'
jinja2==3.1.2 ; python_version >= '3.7'
jsonfield==3.1.0 ; python_version >= '3.6'
jsonpickle==3.0.2 ; python_version >= '3.7'
markdown==3.3.7 ; python_version >= '3.6'
markupsafe==2.1.3 ; python_version >= '3.7'
martor==1.6.28
matplotlib-inline==0.1.6 ; python_version >= '3.5'
networkx==3.2.1
parso==0.8.3 ; python_version >= '3.6'
pipfile==0.0.2
prompt-toolkit==3.0.41 ; python_full_version >= '3.7.0'
pure-eval==0.2.2
pygments==2.17.2 ; python_version >= '3.7'
pytz==2023.3.post1
pyvis==0.3.2
rapidfuzz==3.5.2 ; python_version >= '3.8'
requests==2.31.0 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.5 ; python_version >= '3.8'
sqlparse==0.4.4 ; python_version >= '3.5'
stack-data==0.6.3
swapper==1.3.0
thefuzz==0.20.0
toml==0.10.2 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
traitlets==5.14.0 ; python_version >= '3.8'
tzdata==2023.3 ; sys_platform == 'win32'
urllib3==2.1.0 ; python_version >= '3.8'
wcwidth==0.2.12
webencodings==0.5.1
zipp==3.17.0 ; python_version >= '3.8'
